Добавила модальное окно для заказа из ленты заказов
