+++ src/components/feed-card-images/feed-card-images.module.css
@@ -2,5 +2,5 @@
   position: relative;
   top: -50px;
   left: 265px;
-  z-index: 100000;
+  z-index: 998;
 }
