+++ src/components/feed-list/feed-list.tsx
@@ -20,7 +20,7 @@ export function FeedList() {
           number={v.number}
           createdAt={v.createdAt}
           name={v.name}
-          click={() => navigate(`${v._id}`)}
+          // click={() => navigate(`${v._id}`)}
         />
       ))}
     </div>
