--- src/pages/feed-detail/feed-detail.module.css
@@ -1,37 +1,0 @@
-.row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.amount {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.statusText {
-  color: #00cccc;
-}
-
-.scrollbar {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  overflow: auto;
-  height: auto;
-  max-height: calc(100vh - 60vh);
-}
-
-.grid {
-  display: flex;
-  justify-content: center;
-}
-
-.gridCard {
-  max-height: 620px;
-  min-height: 620px;
-  max-width: 640px;
-  min-width: 640px;
-}
